% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkNames}
\alias{.checkNames}
\title{Missing names check}
\usage{
.checkNames(
  DGEdata = DGEdata,
  x = x,
  y = y,
  ID = ID,
  symbol = symbol,
  logFC = logFC,
  FDR = FDR
)
}
\arguments{
\item{DGEdata}{The object to plot from:
\itemize{
\item \code{limma:} A \link[limma:MArrayLM-class]{MArrayLM} object from
\link[limma]{eBayes} or \link[limma]{treat}
\item \code{edgeR:} A list of \link[edgeR:DGELRT-class]{DGELRT} objects
from \link[edgeR]{glmQLFTest}, \link[edgeR]{glmTreat},
or \link[edgeR]{glmLRT}
\item \code{DESeq2:} a \link[DESeq2:DESeqDataSet-class]{DESeqDataSet} from
\link[DESeq2]{DESeq} or a list of
\link[DESeq2:DESeqResults-class]{DESeqResults} from \link[DESeq2]{results}
\item \code{Other packages:} A list of data.frames,
see details section for more information
}}

\item{x}{Character specifying the name of DGE results within the object
for the x-axis}

\item{y}{Character specifying the name of DGE results within the object
for the y-axis}

\item{ID}{Column name for gene IDs}

\item{symbol}{Column name for gene symbols, which can be the same as the
value for the ID column if not present in the object}

\item{logFC}{Column name for logFC values}

\item{FDR}{Column name for FDR values}
}
\value{
A character
}
\description{
Check for missing names in the DGEdata object
}
\keyword{internal}
