% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.plot4way}
\alias{.plot4way}
\title{gg4way plot}
\usage{
.plot4way(
  DGEtibble = DGEtibble,
  x = x,
  y = y,
  sep = sep,
  logFCcutoff = logFCcutoff,
  lineColor = lineColor,
  colorKey = colorKey,
  corRes = corRes,
  textKey = textKey,
  hjust = hjust,
  vjust = vjust,
  textSize = textSize,
  label = label
)
}
\arguments{
\item{x}{Character specifying the name of DGE results within the object
for the x-axis}

\item{y}{Character specifying the name of DGE results within the object
for the y-axis}

\item{sep}{Character specifying the separator between conditions
for the contrast name provided to the x and y arguments}

\item{logFCcutoff}{Numeric for the absolute logFC cut-off for DEGs
(default is 1)}

\item{lineColor}{Color of lines}

\item{textSize}{Numeric specifying size of text with gene
overlap category totals, where 0 will remove the text}

\item{label}{Character vector specifying the symbols of genes to label
(FALSE for none, TRUE for all blue)}
}
\value{
A \link[ggplot2]{ggplot}
}
\description{
Creates a 4way plot
}
\keyword{internal}
