% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{.prepareAnnotations}
\alias{.prepareAnnotations}
\title{Prepare annotations}
\usage{
.prepareAnnotations(
  countTibble = countTibble,
  colorKey = colorKey,
  textOffset = textOffset
)
}
\arguments{
\item{countTibble}{A \link[tibble:tbl_df-class]{tibble} of summarized counts}

\item{textOffset}{Numeric specifying offset of text with sums}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Prepare text annotations of sums for plotting
}
\keyword{internal}
