% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{.prepareData}
\alias{.prepareData}
\title{Prepare data}
\usage{
.prepareData(
  tidyList = tidyList,
  x = x,
  y = y,
  ID = ID,
  symbol = symbol,
  logFC = logFC,
  FDR = FDR,
  logFCcutoff = logFCcutoff,
  FDRcutoff = FDRcutoff
)
}
\arguments{
\item{tidyList}{A list of tibbles with DGE results from limma or DESeq2}

\item{x}{Character specifying name of DGE results for the x-axis}

\item{y}{Character specifying name of DGE results for the y-axis}

\item{ID}{Column name for gene ID}

\item{symbol}{Column name for gene symbol description}

\item{logFC}{Column name for logFC values}

\item{FDR}{Column name for FDR values}

\item{logFCcutoff}{Numeric for the absolute Log2FC cut-off for DEGs,
default is 1}

\item{FDRcutoff}{Numeric for the FDR cut-off for DEGs, default is 0.05}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Prepare data for a 4way plot
}
\keyword{internal}
