% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{.sumCounts}
\alias{.sumCounts}
\title{Summarize counts}
\usage{
.sumCounts(DGEtibble = DGEtibble, x = x, y = y)
}
\arguments{
\item{DGEtibble}{A \link[tibble:tbl_df-class]{tibble} of DGE results}

\item{x}{Character specifying name of DGE results for the x-axis}

\item{y}{Character specifying name of DGE results for the y-axis}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Create a summary table counts for DGE contrast overlaps for
shared (quadrants) and non-shared (lines) DEGs
}
\keyword{internal}
