% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{.testCor}
\alias{.testCor}
\title{Correlation test}
\usage{
.testCor(DGEtibble = DGEtibble)
}
\arguments{
\item{DGEtibble}{A \link[tibble:tbl_df-class]{tibble} of DGE results}
}
\value{
A numeric of the Pearson correlation
}
\description{
Test the correlation between DGE contrasts
}
\keyword{internal}
