% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tidyLabel}
\alias{.tidyLabel}
\title{Tidy axis labels}
\usage{
.tidyLabel(label = NULL, sep = " vs ", labelType = c("x", "y"))
}
\arguments{
\item{label}{Character vector specifying the symbols of genes to label
(FALSE for none, TRUE for all blue)}

\item{sep}{Character specifying the separator between conditions
for the contrast name provided to the x and y arguments}
}
\value{
A \link[base]{call}
}
\description{
Process axis labels from contrast names
}
\keyword{internal}
