% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.totalCounts}
\alias{.totalCounts}
\title{Summarize counts}
\usage{
.totalCounts(DGEtibble = DGEtibble, x = x, y = y, logFCcutoff = logFCcutoff)
}
\arguments{
\item{DGEtibble}{A \link[tibble:tbl_df-class]{tibble} of DGE results}

\item{x}{Character specifying the name of DGE results within the object
for the x-axis}

\item{y}{Character specifying the name of DGE results within the object
for the y-axis}

\item{logFCcutoff}{Numeric for the absolute logFC cut-off for DEGs
(default is 1)}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Create a summary table counts for DGE contrast overlaps for
shared (quadrants) and non-shared (lines) DEGs
}
\keyword{internal}
