% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyDGE}
\alias{tidyDGE}
\title{Tidy DGE data for plotting}
\usage{
tidyDGE(fit)
}
\arguments{
\item{fit}{A \link[limma:MArrayLM-class]{MArrayLM} object from
\link[limma]{eBayes}, a \link[DESeq2:DESeqDataSet-class]{DESeqDataSet} from
\link[DESeq2]{DESeq}, or a list of
\link[DESeq2:DESeqResults-class]{DESeqResults} from \link[DESeq2]{results}}
}
\value{
A named list of \link[tibble:tbl_df-class]{tibbles}
with at least the following columns
from a \link[limma:MArrayLM-class]{MArrayLM} object:
\itemize{
\item \code{symbol} Gene symbol
\item \code{ID} Gene ID
\item \code{logFC} Log2FC
\item \code{adj.P.Val} Adjusted p-value
}
}
\description{
Tidy DGE data for \link[gg4way]{gg4way}
}
\examples{
data("airwayFit")
airwayFit |>
    tidyDGE()
}
