% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airwayFit}
\alias{airwayFit}
\title{airwayFit data}
\format{
An object of class \code{MArrayLM} with 14516 rows and 2 columns.
}
\usage{
data(airwayFit)
}
\value{
A \link[limma:MArrayLM-class]{MArrayLM}
}
\description{
Generate example data using
the \link[airway]{airway} data package
}
\keyword{datasets}
