% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{.plot4way}
\alias{.plot4way}
\title{gg4way plot}
\usage{
.plot4way(
  DGEtibble = DGEtibble,
  x = x,
  y = y,
  sep = sep,
  logFCcutoff = logFCcutoff,
  lineColor = lineColor,
  colorKey = colorKey,
  corRes = corRes,
  textKey = textKey,
  hjust = hjust,
  vjust = vjust,
  textSize = textSize,
  label = label
)
}
\arguments{
\item{x}{Character specifying name of DGE results for the x-axis}

\item{y}{Character specifying name of DGE results for the y-axis}

\item{sep}{Character specifying the separator between conditions
for the contrast}

\item{logFCcutoff}{Numeric for the absolute Log2FC cut-off for DEGs,
default is 1}

\item{lineColor}{Color of lines}

\item{textSize}{Numeric specifying size of text with sums}

\item{label}{Character vector specifying genes to label
(FALSE for none, "shared" for all blue)}
}
\value{
A \link[ggplot2]{ggplot}
}
\description{
Creates a 4way plot
}
\keyword{internal}
