% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepareAnnotations}
\alias{.prepareAnnotations}
\title{Prepare annotations}
\usage{
.prepareAnnotations(
  totalTibble = totalTibble,
  colorKey = colorKey,
  textNudge = textNudge
)
}
\arguments{
\item{totalTibble}{A \link[tibble:tbl_df-class]{tibble} of summarized counts}

\item{textNudge}{Numeric specifying nudge of text with gene
overlap category totals}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Prepare text annotations of sums for plotting
}
\keyword{internal}
