% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepareData}
\alias{.prepareData}
\title{Prepare data}
\usage{
.prepareData(
  DGEdata = DGEdata,
  x = x,
  y = y,
  ID = ID,
  symbol = symbol,
  logFC = logFC,
  FDR = FDR,
  logFCcutoff = logFCcutoff,
  FDRcutoff = FDRcutoff
)
}
\arguments{
\item{DGEdata}{The object to plot from:
\itemize{
\item \code{limma:} A \link[limma:MArrayLM-class]{MArrayLM} object from
\link[limma]{eBayes} or \link[limma]{treat}
\item \code{edgeR:} A list of \link[edgeR:DGELRT-class]{DGELRT} objects
from \link[edgeR]{glmQLFTest}, \link[edgeR]{glmTreat},
or \link[edgeR]{glmLRT}
\item \code{DESeq2:} a \link[DESeq2:DESeqDataSet-class]{DESeqDataSet} from
\link[DESeq2]{DESeq} or a list of
\link[DESeq2:DESeqResults-class]{DESeqResults} from \link[DESeq2]{results}
\item \code{Other packages:} A list of data.frames,
see details section for more infromation
}}

\item{x}{Character specifying name of DGE results within object
for the x-axis}

\item{y}{Character specifying name of DGE results within object
for the y-axis}

\item{ID}{Column name for gene ID}

\item{symbol}{Column name for gene symbol description}

\item{logFC}{Column name for logFC values}

\item{FDR}{Column name for FDR values}

\item{logFCcutoff}{Numeric for the absolute Log2FC cut-off for DEGs,
default is 1}

\item{FDRcutoff}{Numeric for the FDR cut-off for DEGs, default is 0.05}
}
\value{
A \link[tibble:tbl_df-class]{tibble}
}
\description{
Prepare data for a 4way plot
}
\keyword{internal}
