% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg4way.R
\name{gg4way}
\alias{gg4way}
\title{Create a 4way plot}
\usage{
gg4way(
  tidyList,
  x = NULL,
  y = NULL,
  ID = "ID",
  symbol = "symbol",
  logFC = "logFC",
  FDR = "adj.P.Val",
  sep = " vs ",
  label = FALSE,
  FDRcutoff = 0.05,
  logFCcutoff = 1,
  colorVector = c("grey80", "firebrick", "forestgreen", "mediumblue"),
  lineColor = "grey60",
  sharedOnly = FALSE,
  corOnly = FALSE,
  textSize = 4,
  textOffset = 0.25
)
}
\arguments{
\item{tidyList}{A list of tibbles with DGE results from limma or DESeq2}

\item{x}{Character specifying name of DGE results for the x-axis}

\item{y}{Character specifying name of DGE results for the y-axis}

\item{ID}{Column name for gene ID}

\item{symbol}{Column name for gene symbol description}

\item{logFC}{Column name for logFC values}

\item{FDR}{Column name for FDR values}

\item{sep}{Character specifying the separator between conditions
for the contrast}

\item{label}{Character vector specifying genes to label
(FALSE for none, "shared" for all blue)}

\item{FDRcutoff}{Numeric for the FDR cut-off for DEGs, default is 0.05}

\item{logFCcutoff}{Numeric for the absolute Log2FC cut-off for DEGs,
default is 1}

\item{colorVector}{Character vector of colors in the following order:
"not significant", "significant in x", "significant in y",
"significant in both"}

\item{lineColor}{Color of lines}

\item{sharedOnly}{Logical indicating whether to only return a tibble of
shared DEGs}

\item{corOnly}{Logical indicating whether to only return the
correlation coefficient}

\item{textSize}{Numeric specifying size of text with sums}

\item{textOffset}{Numeric specifying offset of text with sums}
}
\value{
A \link[ggplot2]{ggplot}
(or numeric or \link[tibble:tbl_df-class]{tibble})
}
\description{
4way plots enable a comparison of two contrasts of differential
gene expression. The gg4way package creates 4way plots using the ggplot2
framework and supports popular Bioconductor objects.
The package also provides information about the correlation between
contrasts and significant genes of interest.
}
\examples{
data("airwayFit")
airwayFit |>
    tidyDGE() |>
    gg4way(x = "N61311 vs N052611",
           y = "N061011 vs N052611")
}
